Move the template definitions into the header file
The template definitions have been moved into the header file such
that we would need to include the source files in the wscript of the
different simulations. The wscript of the ospf-network and ppfs
simulator has been updated accordingly.

Files with the .tpp extension contain template function
definitions. These were used to separate code from the class
definition.