Create a static class to handle random number generation
The random number generation functionality has been moved to the
static class RandomGeneratorManager. This class will handle the
creation of the random number generator as well as take care of the
seed and run variables. These seed and run variables are global to the
whole simulation.

The PPFS and OSPF simulations have been updated to use this class to
handle their random number generator functionality.