Move node creation code to the header file
TopologyBuilder is a template class therefore the implementation needs to be in the header file. Previously, the topology-builder.cc file was being included in the header file. This made for obscure code that wasn't obvious. Therefore the implementation code is now being moved to the header file *below* the class definition.