Update LinkContainer to store a pointer to the Link not the object
The LinkContainer map has been updated to store a unique pointer to the Link object instead of the link object itself. This was causing problems when pointers to that object were being stored in the Path class.
Note that all the code used for testing using cout has been removed.