Log the number of dropped packets per switch
Add a ResultsContainer instance to each switch so that we can keep track
of all the packets that were dropped. This change required some drastic
code changes due to the cyclic reference between the SwitchContainer and
the ResultsContainer class. This issue has been solved by using forward
declarations where necessary.