Associate a path id when logging a unipath transmission
A uni-path transmitter will only transmit flows over a single path. If
the flow is assigned only a single path it's all well and good. However,
when used in conjunction with PPFS, it is the PPFS switches that will
take care of splitting the packets over multiple paths.
However, for logging purposes a valid path id needs to be given when
logging transmission because of the packet transmission counter.
Therefore, all packets will be logged as they are transmitted on the
first path id of the given flow.