Fix bug in ReceiveBuffer::RemovePacket
There was an issue in the receiver buffer that was causing an assertion
failure because the function that is used to calculate the packet size
when removing a packet from the buffer includes the P2P header while
the function that adds packets to the receive buffer does not.

This was causing inconsistency in the addition/subtraction of the buffer
usage causing the crash because we were removing more bytes than we were
adding.